+++ packages/sales-api-service/src/services/__tests__/recurring-payments.service.spec.js
@@ -13,17 +13,31 @@ import {
   generateRecurringPaymentRecord,
   processRPResult,
   findNewestExistingRecurringPaymentInCrm,
+  getRecurringPaymentAgreement,
   cancelRecurringPayment
 } from '../recurring-payments.service.js'
 import { calculateEndDate, generatePermissionNumber } from '../permissions.service.js'
 import { getObfuscatedDob } from '../contacts.service.js'
 import { createHash } from 'node:crypto'
-import { AWS } from '@defra-fish/connectors-lib'
+import { AWS, govUkPayApi } from '@defra-fish/connectors-lib'
 import { TRANSACTION_STAGING_TABLE, TRANSACTION_QUEUE } from '../../config.js'
 import { TRANSACTION_STATUS } from '../../services/transactions/constants.js'
 import { retrieveStagedTransaction } from '../../services/transactions/retrieve-transaction.js'
 import { createPaymentJournal, getPaymentJournal, updatePaymentJournal } from '../../services/paymentjournals/payment-journals.service.js'
 import { PAYMENT_JOURNAL_STATUS_CODES, TRANSACTION_SOURCE, PAYMENT_TYPE } from '@defra-fish/business-rules-lib'
+import db from 'debug'
+
+jest.mock('ioredis', () => ({
+  built: {
+    utils: {
+      debug: jest.fn()
+    }
+  }
+}))
+
+jest.mock('debug', () => jest.fn(() => jest.fn()))
+const { value: debug } = db.mock.results[db.mock.calls.findIndex(c => c[0] === 'sales:recurring')]
+
 const { docClient, sqs } = AWS.mock.results[0].value
 
 jest.mock('@defra-fish/dynamics-lib', () => ({
@@ -46,7 +60,10 @@ jest.mock('@defra-fish/connectors-lib', () => ({
     sqs: {
       sendMessage: jest.fn()
     }
-  }))
+  })),
+  govUkPayApi: {
+    getRecurringPaymentAgreementInformation: jest.fn()
+  }
 }))
 
 jest.mock('node:crypto', () => ({
@@ -91,6 +108,7 @@ jest.mock('@defra-fish/business-rules-lib', () => ({
     debit: Symbol('debit')
   }
 }))
+global.structuredClone = obj => JSON.parse(JSON.stringify(obj))
 
 const getMockRecurringPayment = (overrides = {}) => ({
   name: 'Test Name',
@@ -227,7 +245,14 @@ const getMockResponse = () => ({
 })
 
 describe('recurring payments service', () => {
-  const createSimpleSampleTransactionRecord = () => ({ payment: { recurring: true }, permissions: [{}] })
+  const createSimpleSampleTransactionRecord = () => ({
+    payment: {
+      recurring: {
+        nextDueDate: '2025-01-01T00:00:00.000Z'
+      }
+    },
+    permissions: [{}]
+  })
   const createSamplePermission = overrides => {
     const p = new Permission()
     p.referenceNumber = 'ABC123'
@@ -247,6 +272,11 @@ describe('recurring payments service', () => {
   beforeAll(() => {
     TRANSACTION_QUEUE.Url = 'TestUrl'
     TRANSACTION_STAGING_TABLE.TableName = 'TestTable'
+    const mockResponse = {
+      ok: true,
+      json: jest.fn().mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: '1234' } } })
+    }
+    govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
   })
 
   describe('getRecurringPayments', () => {
@@ -284,13 +314,13 @@ describe('recurring payments service', () => {
       const transactionRecord = {
         payment: {
           recurring: {
-            name: 'Test Name',
-            nextDueDate: new Date('2023-11-02'),
+            nextDueDate: '2023-11-02T00:00:00.000Z',
             cancelledDate: null,
             cancelledReason: null,
-            endDate: new Date('2023-11-12'),
+            endDate: '2023-11-12T00:00:00.000Z',
             agreementId: '435678',
-            status: 0
+            status: 0,
+            last_digits_card_number: '0128'
           }
         },
         permissions: [getMockPermission()]
@@ -298,6 +328,19 @@ describe('recurring payments service', () => {
       const contact = getMockContact()
       const result = await processRecurringPayment(transactionRecord, contact)
       expect(result.recurringPayment).toMatchSnapshot()
+    })
+
+    it('should set a valid name on the recurringPayment', async () => {
+      const transactionRecord = {
+        payment: {
+          recurring: {
+            nextDueDate: '2023-07-07T00:00:00.000Z'
+          }
+        },
+        permissions: [getMockPermission()]
+      }
+      const result = await processRecurringPayment(transactionRecord, getMockContact())
+      expect(result.recurringPayment.name).toBe('Fester Tester 2023')
     })
 
     it.each(['abc-123', 'def-987'])('generates a publicId %s for the recurring payment', async samplePublicId => {
@@ -336,7 +379,7 @@ describe('recurring payments service', () => {
   })
 
   describe('generateRecurringPaymentRecord', () => {
-    const createFinalisedSampleTransaction = (agreementId, permission) => ({
+    const createFinalisedSampleTransaction = (agreementId, permission, lastDigitsCardNumbers) => ({
       expires: 1732892402,
       cost: 35.8,
       isRecurringPaymentSupported: true,
@@ -359,7 +402,8 @@ describe('recurring payments service', () => {
       id: 'd26d646f-ed0f-4cf1-b6c1-ccfbbd611757',
       dataSource: 'Web Sales',
       transactionId: 'd26d646f-ed0f-4cf1-b6c1-ccfbbd611757',
-      status: { id: 'FINALISED' }
+      status: { id: 'FINALISED' },
+      lastDigitsCardNumbers
     })
 
     it.each([
@@ -371,7 +415,8 @@ describe('recurring payments service', () => {
           issueDate: '2024-11-22T15:00:45.922Z',
           endDate: '2025-11-21T23:59:59.999Z'
         },
-        '2025-11-12T00:00:00.000Z'
+        '2025-11-12T00:00:00.000Z',
+        '1234'
       ],
       [
         'next day start - next due on end date minus ten days',
@@ -381,7 +426,8 @@ describe('recurring payments service', () => {
           issueDate: '2024-11-22T15:00:45.922Z',
           endDate: '2025-11-22T23:59:59.999Z'
         },
-        '2025-11-12T00:00:00.000Z'
+        '2025-11-12T00:00:00.000Z',
+        '5678'
       ],
       [
         'starts ten days after issue - next due on issue date plus one year',
@@ -391,7 +437,8 @@ describe('recurring payments service', () => {
           issueDate: '2024-11-12T15:00:45.922Z',
           endDate: '2025-11-21T23:59:59.999Z'
         },
-        '2025-11-12T00:00:00.000Z'
+        '2025-11-12T00:00:00.000Z',
+        '9012'
       ],
       [
         'starts twenty days after issue - next due on issue date plus one year',
@@ -401,7 +448,8 @@ describe('recurring payments service', () => {
           issueDate: '2024-11-12T15:00:45.922Z',
           endDate: '2025-01-30T23:59:59.999Z'
         },
-        '2025-11-12T00:00:00.000Z'
+        '2025-11-12T00:00:00.000Z',
+        '3456'
       ],
       [
         "issued on 29th Feb '24, starts on 30th March '24 - next due on 28th Feb '25",
@@ -411,7 +459,8 @@ describe('recurring payments service', () => {
           issueDate: '2024-02-29T12:38:24.123Z',
           endDate: '2025-03-29T23:59:59.999Z'
         },
-        '2025-02-28T00:00:00.000Z'
+        '2025-02-28T00:00:00.000Z',
+        '7890'
       ],
       [
         "issued on 30th March '25 at 1am, starts at 1:30am - next due on 20th March '26",
@@ -421,13 +470,21 @@ describe('recurring payments service', () => {
           issueDate: '2025-03-30T01:00:00.000Z',
           endDate: '2026-03-29T23:59:59.999Z'
         },
-        '2026-03-20T00:00:00.000Z'
+        '2026-03-20T00:00:00.000Z',
+        '1199'
       ]
-    ])('creates record from transaction with %s', (_d, agreementId, permissionData, expectedNextDueDate) => {
-      const sampleTransaction = createFinalisedSampleTransaction(agreementId, permissionData)
+    ])('creates record from transaction with %s', async (_d, agreementId, permissionData, expectedNextDueDate, lastDigitsCardNumbers) => {
+      const mockResponse = {
+        ok: true,
+        json: jest
+          .fn()
+          .mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: lastDigitsCardNumbers } } })
+      }
+      govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
+      const sampleTransaction = createFinalisedSampleTransaction(agreementId, permissionData, lastDigitsCardNumbers)
       const permission = createSamplePermission(permissionData)
 
-      const rpRecord = generateRecurringPaymentRecord(sampleTransaction, permission)
+      const rpRecord = await generateRecurringPaymentRecord(sampleTransaction, permission)
 
       expect(rpRecord).toEqual(
         expect.objectContaining({
@@ -439,7 +496,8 @@ describe('recurring payments service', () => {
               cancelledReason: null,
               endDate: permissionData.endDate,
               agreementId,
-              status: 1
+              status: 1,
+              last_digits_card_number: lastDigitsCardNumbers
             })
           }),
           permissions: expect.arrayContaining([permission])
@@ -464,22 +522,26 @@ describe('recurring payments service', () => {
           endDate: '2025-11-10T23:59:59.999Z'
         }
       ]
-    ])('throws an error for invalid dates when %s', (_d, permission) => {
+    ])('throws an error for invalid dates when %s', async (_d, permission) => {
       const sampleTransaction = createFinalisedSampleTransaction('hyu78ijhyu78ijuhyu78ij9iu6', permission)
 
-      expect(() => generateRecurringPaymentRecord(sampleTransaction)).toThrow('Invalid dates provided for permission')
+      await expect(generateRecurringPaymentRecord(sampleTransaction)).rejects.toThrow('Invalid dates provided for permission')
     })
 
-    it('returns a false flag when agreementId is not present', () => {
-      const sampleTransaction = createFinalisedSampleTransaction(null, {
-        startDate: '2024-11-22T15:30:45.922Z',
-        issueDate: '2024-11-22T15:00:45.922Z',
-        endDate: '2025-11-21T23:59:59.999Z'
-      })
+    it('returns a false flag when agreementId is not present', async () => {
+      const sampleTransaction = createFinalisedSampleTransaction(
+        null,
+        {
+          startDate: '2024-11-22T15:30:45.922Z',
+          issueDate: '2024-11-22T15:00:45.922Z',
+          endDate: '2025-11-21T23:59:59.999Z'
+        },
+        '0123'
+      )
 
-      const rpRecord = generateRecurringPaymentRecord(sampleTransaction)
+      const rpRecord = await generateRecurringPaymentRecord(sampleTransaction)
 
-      expect(rpRecord.payment.recurring).toBeFalsy()
+      expect(rpRecord.payment?.recurring).toBeFalsy()
     })
   })
 
@@ -734,6 +796,64 @@ describe('recurring payments service', () => {
       dynamicsClient.retrieveMultipleRequest.mockReturnValueOnce({ value: [] })
       const rcp = await findNewestExistingRecurringPaymentInCrm()
       expect(rcp).toBeFalsy()
+    })
+  })
+
+  describe('getRecurringPaymentAgreement', () => {
+    const agreementId = '1234'
+
+    it('should send provided agreement id data to Gov.UK Pay', async () => {
+      const mockResponse = {
+        ok: true,
+        json: jest.fn().mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: '1234' } } })
+      }
+      govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
+      await getRecurringPaymentAgreement(agreementId)
+      expect(govUkPayApi.getRecurringPaymentAgreementInformation).toHaveBeenCalledWith(agreementId)
+    })
+
+    it('should return response body when payment creation is successful', async () => {
+      const mockResponse = {
+        ok: true,
+        json: jest.fn().mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: '1234' } } })
+      }
+      govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
+
+      const result = await getRecurringPaymentAgreement(agreementId)
+
+      expect(result).toEqual({
+        success: true,
+        payment_instrument: {
+          card_details: {
+            last_digits_card_number: '1234'
+          }
+        }
+      })
+    })
+
+    it('debug should output message when response.ok is true without card details', async () => {
+      const mockResponse = {
+        ok: true,
+        json: jest.fn().mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: '1234' } } })
+      }
+      govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
+
+      await getRecurringPaymentAgreement(agreementId)
+
+      expect(debug).toHaveBeenCalledWith('Successfully got recurring payment agreement information: %o', {
+        success: true,
+        payment_instrument: {}
+      })
+    })
+
+    it('should throw an error when the response is not ok', async () => {
+      const mockResponse = {
+        ok: false,
+        json: jest.fn().mockResolvedValue({ success: true, payment_instrument: { card_details: { last_digits_card_number: '1234' } } })
+      }
+      govUkPayApi.getRecurringPaymentAgreementInformation.mockResolvedValue(mockResponse)
+
+      await expect(getRecurringPaymentAgreement(agreementId)).rejects.toThrow('Failure getting agreement in the GOV.UK API service')
     })
   })