+++ CHANGELOG.md
@@ -1,4 +1,49 @@
 
+## v1.62.0-rc.8 (2025-07-23)
+
+#### :rocket: Enhancement
+* `connectors-lib`, `sales-api-service`
+  * [#2145](https://github.com/DEFRA/rod-licensing/pull/2145) Store bank card digits ([@ScottDormand96](https://github.com/ScottDormand96))
+
+#### Committers: 1
+- Scott Dormand ([@ScottDormand96](https://github.com/ScottDormand96))
+
+## v1.62.0-rc.7 (2025-07-22)
+
+#### :rocket: Enhancement
+* `business-rules-lib`, `recurring-payments-job`
+  * [#2148](https://github.com/DEFRA/rod-licensing/pull/2148) Recurring payments failed payments ([@ScottDormand96](https://github.com/ScottDormand96))
+
+#### Committers: 1
+- Scott Dormand ([@ScottDormand96](https://github.com/ScottDormand96))
+
+## v1.62.0-rc.6 (2025-07-21)
+
+#### :bug: Bug Fix
+* `gafl-webapp-service`
+  * [#2180](https://github.com/DEFRA/rod-licensing/pull/2180) Use blue favicon ([@irisfaraway](https://github.com/irisfaraway))
+
+#### Committers: 1
+- Iris Faraway ([@irisfaraway](https://github.com/irisfaraway))
+
+## v1.62.0-rc.5 (2025-07-21)
+
+#### :bug: Bug Fix
+* `gafl-webapp-service`
+  * [#2178](https://github.com/DEFRA/rod-licensing/pull/2178) Sanitise payment details in logging ([@irisfaraway](https://github.com/irisfaraway))
+
+#### Committers: 1
+- Iris Faraway ([@irisfaraway](https://github.com/irisfaraway))
+
+## v1.62.0-rc.4 (2025-07-21)
+
+#### :rocket: Enhancement
+* `sales-api-service`
+  * [#2179](https://github.com/DEFRA/rod-licensing/pull/2179) Set name of RecurringPayment record ([@irisfaraway](https://github.com/irisfaraway))
+
+#### Committers: 1
+- Iris Faraway ([@irisfaraway](https://github.com/irisfaraway))
+
 ## v1.62.0-rc.3 (2025-07-10)
 
 #### :rocket: Enhancement